Fix [#104] 삭제 또는 추가 파일이 전달되지 않을 경우 처리

+++ src/main/java/org/cotato/tlinkserver/api/facade/HomeworkFacade.java
@@ -73,17 +73,21 @@ public class HomeworkFacade {
 		Homework homework = homeworkService.getHomework(homeworkId);
 		List<HomeworkFile> homeworkFiles = homework.getHomeworkFiles();
 
-		removeHomeworkFiles.forEach(id -> {
-			HomeworkFile homeworkFile = homeworkFiles.stream()
-				.filter(file -> file.getId().equals(id))
-				.findFirst()
-				.orElseThrow(() -> new NotFoundException(ErrorMessage.NOT_FOUND));
+		if (removeHomeworkFiles != null) {
+			removeHomeworkFiles.forEach(id -> {
+				HomeworkFile homeworkFile = homeworkFiles.stream()
+					.filter(file -> file.getId().equals(id))
+					.findFirst()
+					.orElseThrow(() -> new NotFoundException(ErrorMessage.NOT_FOUND));
 
-			s3FileHandler.deleteFile(DIRECTORY_PATH + homeworkFile.getS3Key());
-			homeworkFiles.remove(homeworkFile);
-		});
+				s3FileHandler.deleteFile(DIRECTORY_PATH + homeworkFile.getS3Key());
+				homeworkFiles.remove(homeworkFile);
+			});
+		}
 
-		this.saveHomeworkFiles(addHomeworkFiles, homework, user);
+		if (addHomeworkFiles != null) {
+			this.saveHomeworkFiles(addHomeworkFiles, homework, user);
+		}
 
 		if (user.getRole().equals(Role.TEACHER)) {
 			homework.setName(homeworkName);
